pyproject.toml
[project]
name = "yolonas"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "i307132", email = "luna.yin@sap.com"},
]
dependencies = [
    "super-gradients==3.2.0",
    "torchvision==0.17.2",
    "imutils>=0.5.4",
    "roboflow>=1.1.36",
    "pytube>=15.0.0",
    "IPython>=8.18.1",
    "supervision>=0.22.0",
    "pytorch-quantization==2.1.3",
    "matplotlib>=3.9.1",
]
requires-python = "<3.11, >=3.9"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
